Periodically run
SetThreadExecutionState(ES_DISPLAY_REQUIRED|ES_SYSTEM_REQUIRED);

to reset the input timer.

Note that this does not use g_application.m_AppFocused, due to a bug: if
the application doesn't have focus to begin with, m_AppFocused is true
until we gain focus and lose it again, because no WM_KILLFOCUS is ever
received.  I'm not fixing that here to keep the scope of this change
small.